separate pca methods from analysis
